@@ -15,58 +15,58 @@
   "keywords": [],
   "license": "MIT",
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-hot-loader": "4.13.0",
+    "react-router-dom": "^6.22.1",
     "styled-components": "5.3.0"
   },
   "devDependencies": {
-    "@babel/core": "7.14.6",
-    "@babel/plugin-transform-runtime": "7.14.5",
-    "@babel/preset-env": "7.14.5",
-    "@babel/preset-react": "7.14.5",
-    "@babel/preset-typescript": "7.14.5",
-    "@electron-forge/cli": "6.0.0-beta.57",
-    "@electron-forge/maker-deb": "6.0.0-beta.57",
-    "@electron-forge/maker-rpm": "6.0.0-beta.57",
-    "@electron-forge/maker-squirrel": "6.0.0-beta.57",
-    "@electron-forge/maker-zip": "6.0.0-beta.57",
-    "@electron-forge/plugin-webpack": "6.0.0-beta.57",
-    "@marshallofsound/webpack-asset-relocator-loader": "0.5.0",
-    "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "11.2.7",
-    "@types/electron-devtools-installer": "2.2.0",
-    "@types/jest": "26.0.23",
-    "@types/react": "17.0.11",
-    "@types/react-dom": "17.0.8",
-    "@types/styled-components": "5.1.10",
-    "@typescript-eslint/eslint-plugin": "4.27.0",
-    "@typescript-eslint/parser": "4.27.0",
-    "autoprefixer": "^10.4.22",
-    "babel-loader": "8.2.2",
-    "cross-env": "7.0.3",
-    "css-loader": "^7.1.2",
-    "electron": "13.1.2",
-    "eslint": "7.29.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-config-standard": "16.0.3",
-    "eslint-plugin-import": "2.23.4",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "3.4.0",
-    "eslint-plugin-promise": "5.1.0",
-    "eslint-plugin-react": "7.24.0",
-    "eslint-plugin-standard": "5.0.0",
+    "@babel/core": "^7.23.0",
+    "@babel/plugin-transform-runtime": "^7.23.0",
+    "@babel/preset-env": "^7.23.0",
+    "@babel/preset-react": "^7.22.0",
+    "@babel/preset-typescript": "^7.23.0",
+    "@electron-forge/cli": "^7.0.0",
+    "@electron-forge/maker-deb": "^7.0.0",
+    "@electron-forge/maker-rpm": "^7.0.0",
+    "@electron-forge/maker-squirrel": "^7.0.0",
+    "@electron-forge/maker-zip": "^7.0.0",
+    "@electron-forge/plugin-webpack": "^7.0.0",
+    "@marshallofsound/webpack-asset-relocator-loader": "^0.5.0",
+    "@testing-library/jest-dom": "^6.1.0",
+    "@testing-library/react": "^13.4.0",
+    "@types/jest": "^29.5.0",
+    "@types/react": "^18.3.1",
+    "@types/react-dom": "^18.3.1",
+    "@types/styled-components": "^5.1.0",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "autoprefixer": "^10.4.16",
+    "babel-loader": "^9.1.0",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.8.0",
+    "electron": "^28.0.0",
+    "eslint": "^8.45.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.28.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-promise": "^6.1.0",
+    "eslint-plugin-react": "^7.33.0",
+    "eslint-plugin-standard": "^5.0.0",
     "file-loader": "^6.2.0",
-    "jest": "27.0.4",
-    "npm-run-all": "4.1.5",
-    "postcss": "^8.5.6",
-    "postcss-loader": "^8.1.1",
-    "prettier": "2.3.1",
-    "style-loader": "^4.0.0",
-    "tailwindcss": "^3.4.18",
-    "ts-jest": "27.0.3",
-    "typescript": "4.3.4",
-    "wait-on": "5.3.0"
+    "jest": "^29.6.0",
+    "npm-run-all": "^4.1.5",
+    "postcss": "^8.4.29",
+    "postcss-loader": "^7.3.0",
+    "prettier": "^3.0.0",
+    "style-loader": "^3.3.0",
+    "tailwindcss": "^3.3.0",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.2.0",
+    "wait-on": "^7.2.0"
   },
   "config": {
     "forge": {
