@@ -1,5 +1,11 @@
-import ReactDOM from "react-dom";
-import { App } from "./App";
+import { createRoot } from 'react-dom/client';
+import { App } from './App';
 import "./index.css";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById('root');
+if (!container) {
+  throw new Error('Root element not found');
+}
+
+const root = createRoot(container);
+root.render(<App />);
